changed the structure in the components folder
