+++ src/components/AppRouter.jsx
@@ -1,27 +1,26 @@
 import React from "react";
-import { Route, Router, Routes } from "react-router-dom";
-import Login from "./Login";
-import MapPage from './MapPage';
-import Profile from "./Profile";
+import { Route, Routes } from "react-router-dom";
+import Login from "./Login/Login";
+import Main from './Main/Main';
 import { RequireAuth } from "./RequireAuth"
 
 const AppRouter = () => {
 
     return (
         <Routes>
-            <Route path="/login" element={<Login formType="login"/>} />
-            <Route path="/registration" element={<Login formType="registration"/>} />
+            <Route path="/login" element={<Login formType="login" />} />
+            <Route path="/registration" element={<Login formType="registration" />} />
             <Route path="/order" element={
                 <RequireAuth>
-                    <MapPage pageType="order"/>
+                    <Main pageType="order" />
                 </RequireAuth>
             } />
             <Route path="/profile" element={
                 <RequireAuth>
-                    <MapPage pageType="profile"/>
+                    <Main pageType="profile" />
                 </RequireAuth>
             } />
-            <Route path="*" element={<Login formType="login"/>} />
+            <Route path="*" element={<Login formType="login" />} />
         </Routes>
     )
 }
