+++ src/components/Login/RegistrationForm.spec.jsx
