+++ src/components/Main/Map.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef } from "react";
 import { connect } from "react-redux";
 import mapboxgl from 'mapbox-gl';
-import '../styles/Map.css'
-import { drawRoute } from "../asyncActions/api"
+import './Map.css'
+import { drawRoute } from "../../asyncActions/api"
 
 mapboxgl.accessToken = 'pk.eyJ1Ijoic3Ryb2xzZSIsImEiOiJjbDNkOG5nbWowMGQyM29sZnNqdW43ZTY5In0.FoQyIdGOq1qDGV7B6mxcjQ';
 
@@ -16,7 +16,7 @@ const Map = ({ isOrdered, coords }) => {
             center: [30.312786581010016, 59.93470138710765],
             zoom: 11
         });
-  
+
         map.on('load', () => {
             if (isOrdered) {
 
@@ -27,9 +27,9 @@ const Map = ({ isOrdered, coords }) => {
         return () => map.remove();
     }, [coords])
 
-    
+
     return (
-        
+
         <div className="map" data-testid="map" ref={mapContainerRef}>
         </div>
     )
