+++ src/components/Main/Profile.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { connect } from "react-redux";
-import { Input, Button, InputLabel, Modal, Typography, Grid, FormControl, FormHelperText, Box, Paper } from "@mui/material";
-import { saveCardAction } from "../redux/actions";
-import ProfileForm from "./ProfileForm";
+import { Grid, Paper } from "@mui/material";
+import { saveCardAction } from "../../redux/actions";
+import ProfileForm from "./ProfileForm/ProfileForm";
 
 const Profile = () => {
 
@@ -12,7 +12,8 @@ const Profile = () => {
             position: "absolute", bottom: 0, top: "110px",
             display: "flex", justifyContent: "center", alignItems: 'center'
         }}>
-            <Grid item component={Paper} xs={12} sm={6} md={6}>
+            <Grid item component={Paper} xs={12} sm={6} md={6}
+                sx={{borderRadius:"10px"}}>
                 <ProfileForm />
             </Grid>
         </Grid>)
